Code cleanup after PR self-review.
